split all code from qpi.h into new sub headers
